feat: add monitoring sample and fix common package.json
