+++ packages/common/src/index.ts
@@ -1,2 +1,3 @@
 export * from './monitoring';
 export * from './interceptors';
+export * from './controller';
