+++ sample/02-monitoring/.prettierrc
