+++ sample/02-monitoring/monitor/grafana/provisioning/dashboards/dashboard.yml
