+++ sample/02-monitoring/monitor/grafana/provisioning/dashboards/docker_containers.json
