+++ sample/02-monitoring/monitor/grafana/provisioning/datasources/datasources.yml
