+++ sample/02-monitoring/monitor/response_times_base.json
