+++ sample/02-monitoring/monitor/setup_monitoring_services.yml
