+++ sample/02-monitoring/nest-cli.json
