+++ sample/02-monitoring/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "package-test",
+  "name": "02-monitoring",
   "version": "0.0.1",
   "description": "",
   "author": "",
@@ -23,8 +23,7 @@
     "@nestjs/common": "^10.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
-    "@techsavvyash/nestjs-monitor": "^0.0.4",
-    "@willsoto/nestjs-prometheus": "^6.0.0",
+    "@samagra-x/stencil": "^0.0.2",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1"
   },
