+++ sample/02-monitoring/test/jest-e2e.json
