+++ sample/02-monitoring/tsconfig.build.json
