+++ sample/02-monitoring/tsconfig.json
